enmacs_controller/kis_ems/requirements.txt
aiohttp
holidays
pyyaml
requests
schedule
tzdata
websockets
